# openapi/expense-sharing-api.yaml
openapi: 3.0.0

info:
  title: Expense Sharing API
  version: 1.0.0

paths:
  /groups:
    post:
      summary: Create Group

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
            examples:
              createGroup:
                value:
                  name: Goa Trip

              validationError:
                value:
                  name: 123

      responses:
        "201":
          description: Group Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
              examples:
                createGroup:
                  value:
                    groupId: 1
                    name: Goa Trip
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  value:
                    detail: Validation error

  /groups/{groupId}:
    get:
      summary: Get Group

      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer

          examples:
            success:
              value: 1

            notFound:
              value: 99999

            validationError:
              value: abc

      responses:
        "200":
          description: Group Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
              examples:
                success:
                  value:
                    groupId: 1
                    name: Goa Trip

        "404":
          description: Group Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    detail: Group not found

        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  value:
                    detail: Validation error

  /expenses:
    post:
      summary: Add Expense

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddExpenseRequest"
            examples:
              addExpense:
                value:
                  groupId: 1
                  amount: 1500.50
                  paidBy: 101

              validationError:
                value:
                  groupId: "abc"
                  amount: 1500.5
                  paidBy: 101

      responses:
        "201":
          description: Expense Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
              examples:
                addExpense:
                  value:
                    expenseId: 1
                    status: CREATED
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  value:
                    detail: Validation error

    get:
      summary: Get Expenses

      responses:
        "200":
          description: Expense List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
              examples:
                expenseList:
                  value:
                    - expenseId: 1
                      groupId: 1
                      amount: 1500.50
                      paidBy: 101

  /balances/{userId}:
    get:
      summary: View Balance

      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          examples:
            balance:
              value: 101

            validationError:
              value: abc

      responses:
        "200":
          description: Balance Details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
              examples:
                balance:
                  value:
                    userId: 101
                    totalOwed: 500
                    totalReceivable: 1200

        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  value:
                    detail: Validation error

  /settlements:
    post:
      summary: Create Settlement

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementRequest"
            examples:
              settlement:
                value:
                  fromUser: 101
                  toUser: 102
                  amount: 500

              validationError:
                value:
                  fromUser: "abc"
                  toUser: 102
                  amount: 500

      responses:
        "201":
          description: Settlement Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementResponse"
              examples:
                settlement:
                  value:
                    settlementId: 1
                    status: SUCCESS
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  value:
                    detail: Validation error

  /actuator/mappings:
    get:
      summary: List registered endpoints

      responses:
        "200":
          description: List of registered API endpoints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActuatorMappingsResponse"

              examples:
                mappings:
                  value:
                    application: expense-sharing-api
                    framework: FastAPI
                    totalEndpoints: 7
                    mappings:
                      - path: /actuator/mappings
                        methods:
                          - GET
                        name: actuator_mappings
                        summary: List all registered API endpoints
                        operationId: actuatorMappings

                      - path: /groups
                        methods:
                          - POST
                        name: create_group
                        summary: Create Group
                        operationId: createGroup

                      - path: /groups/{groupId}
                        methods:
                          - GET
                        name: get_group
                        summary: Get Group
                        operationId: getGroup

                      - path: /expenses
                        methods:
                          - POST
                        name: add_expense
                        summary: Add Expense
                        operationId: addExpense

                      - path: /expenses
                        methods:
                          - GET
                        name: get_expenses
                        summary: Get Expenses
                        operationId: getExpenses

                      - path: /balances/{userId}
                        methods:
                          - GET
                        name: get_balance
                        summary: View Balance
                        operationId: getBalance

                      - path: /settlements
                        methods:
                          - POST
                        name: create_settlement
                        summary: Create Settlement
                        operationId: createSettlement

components:
  schemas:
    CreateGroupRequest:
      type: object
      required:
        - name

      properties:
        name:
          type: string
          example: Goa Trip

    GroupResponse:
      type: object

      properties:
        groupId:
          type: integer
          example: 1

        name:
          type: string
          example: Goa Trip

    AddExpenseRequest:
      type: object

      required:
        - groupId
        - amount
        - paidBy

      properties:
        groupId:
          type: integer
          example: 1

        amount:
          type: number
          example: 1500.50

        paidBy:
          type: integer
          example: 101

    ExpenseResponse:
      type: object

      properties:
        expenseId:
          type: integer
          example: 1

        status:
          type: string
          example: CREATED

    BalanceResponse:
      type: object

      properties:
        userId:
          type: integer
          example: 101

        totalOwed:
          type: number
          example: 500.00

        totalReceivable:
          type: number
          example: 1200.00

    SettlementRequest:
      type: object

      required:
        - fromUser
        - toUser
        - amount

      properties:
        fromUser:
          type: integer
          example: 101

        toUser:
          type: integer
          example: 102

        amount:
          type: number
          example: 500.00

    SettlementResponse:
      type: object

      properties:
        settlementId:
          type: integer
          example: 1

        status:
          type: string
          example: SUCCESS

    ErrorResponse:
      type: object

      properties:
        detail:
          type: string

      example:
        detail: Group not found

    Expense:
      type: object

      properties:
        expenseId:
          type: integer

        groupId:
          type: integer

        amount:
          type: number

        paidBy:
          type: integer

    ActuatorMapping:
      type: object

      required:
        - path
        - methods
        - name
        - summary
        - operationId

      properties:
        path:
          type: string
        methods:
          type: array
          items:
            type: string
        name:
          type: string
        summary:
          type: string
          nullable: true
        operationId:
          type: string
          nullable: true

    ActuatorMappingsResponse:
      type: object

      required:
        - application
        - framework
        - totalEndpoints
        - mappings

      properties:
        application:
          type: string
        framework:
          type: string
        totalEndpoints:
          type: integer
        mappings:
          type: array
          items:
            $ref: "#/components/schemas/ActuatorMapping"